make the install script more "secure"

+++ stupid-install-script.sh
@@ -1,117 +1,110 @@
+
 #!/usr/bin/env bash
+
+# Function to create a symbolic link but ask for backup if it already exists
+create_symlink() {
+	local target="$1"
+	local link_name="$2"
+
+	if [ -e "$link_name" ]; then
+		read -p "File $link_name already exists. Do you want to backup it? (y/n): " backup_choice
+		if [[ $backup_choice == "y" ]]; then
+			mv "$link_name" "${link_name}.bak"
+			echo "Backup created: ${link_name}.bak"
+		else
+			echo "Skipping creation of symlink for $link_name."
+			return
+		fi
+	fi
+
+	ln -sT "$target" "$link_name"
+	echo "Created symlink: $link_name -> $target"
+}
 
 # System
 sudo cp ~/dotfiles/system/sudoers/florian /etc/sudoers.d/florian
 
-
-
 # Audio
 sudo zypper install sof-firmware
 
-
-
 # Bashrc and Profile
-ln -sT ~/dotfiles/shell/opensuse/.bashrc ~/.bashrc-default
+create_symlink ~/dotfiles/shell/opensuse/.bashrc ~/.bashrc-default
 if ! grep -q "source ~/.bashrc-default" ~/.bashrc; then
   echo "source ~/.bashrc-default" >> ~/.bashrc
 fi
-ln -sT ~/dotfiles/shell/opensuse/.profile ~/.profile-default
+create_symlink ~/dotfiles/shell/opensuse/.profile ~/.profile-default
 if ! grep -q "source ~/.profile-default" ~/.profile; then
   echo "source ~/.profile-default" >> ~/.profile
 fi
-
-
 
 # General
 # sudo zypper install eza
 # mkdir -p ~/.config/eza
 # ln -sT ~/dotfiles/eza/rose-pine.yml  ~/.config/eza/theme.yml
+
 sudo zypper install bat
 mkdir -p "$(bat --config-dir)"
-ln -sT ~/dotfiles/bat/themes "$(bat --config-dir)/themes"
-ln -sT ~/dotfiles/bat/config "$(bat --config-dir)/config"
+create_symlink ~/dotfiles/bat "$(bat --config-dir)"
 bat cache --build
 
 sudo zypper install fd
 
-
-
 # Hyprland
 sudo zypper install hyprland hyprpaper hyprcursor hyprlock hypridle hyprland-qtutils hyprland-qt-support
-## TODO: perhaps use the e.g. pkgconfig(pixman-1) syntax
-sudo zypper install cpio cmake git meson gcc gcc-c++
-sudo zypper install hyprland-devel hyprland-protocols-devel
-sudo zypper install libpixman-1-0-devel libdrm-devel pango-devel libinput-devel wayland-devel wayland-protocols-devel libxkbcommon-devel systemd-devel Mesa-libGLESv2-devel Mesa-libGLESv3-devel hyprlang-devel hyprutils-devel hyprgraphics-devel aquamarine-devel xcb-util-wm-devel
-## Ask plugins version 
-read -p "Enter the version of hyprland-plugins (e.g. v0.48.0): " plugins_version
-hyprpm add https://github.com/hyprwm/hyprland-plugins $plugins_version
-ln -sT ~/dotfiles/hypr ~/.config/hypr
+read -p "Do you want to install the Hyprland plugins ? (y/n): " install_hyprland_devel
+if [[ $install_hyprland_devel == "y" ]]; then
+  # TODO: perhaps use the e.g. pkgconfig(pixman-1) syntax
+  sudo zypper install cpio cmake git meson gcc gcc-c++
+  sudo zypper install hyprland-devel hyprland-protocols-devel
+  sudo zypper install libpixman-1-0-devel libdrm-devel pango-devel libinput-devel wayland-devel wayland-protocols-devel libxkbcommon-devel systemd-devel Mesa-libGLESv2-devel Mesa-libGLESv3-devel hyprlang-devel hyprutils-devel hyprgraphics-devel aquamarine-devel xcb-util-wm-devel
+  read -p "Enter the version of hyprland-plugins (e.g. v0.48.0): " plugins_version
+  hyprpm add https://github.com/hyprwm/hyprland-plugins $plugins_version
+  hyprpm enable hyprexpo
+fi
 sudo zypper install xdg-desktop-portal-hyprland xdg-desktop-portal-gtk
 sudo zypper in qt6-wayland libqt5-qtwayland
-
-
-
+create_symlink ~/dotfiles/hypr ~/.config/hypr
 
 # System controls
 sudo zypper in playerctl brightnessctl
 sudo zypper in power-profiles-daemon
 
-
-
 # Waybar
 sudo zypper install --no-recommends waybar
 ## Recommended
 sudo zypper install upower
-ln -sT ~/dotfiles/waybar ~/.config/waybar
-
-
+create_symlink ~/dotfiles/waybar ~/.config/waybar
 
 # Kvantum
 sudo zypper install kvantum-manager kvantum-qt5 kvantum-qt6
-ln -sT ~/dotfiles/kvantum ~/.config/Kvantum
-
-
-
+create_symlink ~/dotfiles/kvantum ~/.config/Kvantum
 
 # Gtk look
 sudo zypper install nwg-look
 
-
-
-
 # Wlogout
 sudo zypper install wlogout
-ln -sT ~/dotfiles/wlogout ~/.config/wlogout
-
-
+create_symlink ~/dotfiles/wlogout ~/.config/wlogout
 
 # Swaync
 sudo zypper install SwayNotificationCenter
-ln -sT ~/dotfiles/swaync ~/.config/swaync
-
-
+create_symlink ~/dotfiles/swaync ~/.config/swaync
 
 # Wofi
 sudo zypper install wofi
-ln -sT ~/dotfiles/wofi ~/.config/wofi
-ln -sT ~/dotfiles/bin/wofi-emoji ~/.local/bin/wofi-emoji
-
+create_symlink ~/dotfiles/wofi ~/.config/wofi
+create_symlink ~/dotfiles/bin/wofi-emoji ~/.local/bin/wofi-emoji
 
 # Alacritty
 sudo zypper install alacritty
-ln -sT ~/dotfiles/alacritty ~/.config/alacritty
-
-
+create_symlink ~/dotfiles/alacritty ~/.config/alacritty
 
 # Starship
 sudo zypper install starship
-ln -sT ~/dotfiles/starship/starship.toml ~/.config/starship.toml
-
-
+create_symlink ~/dotfiles/starship/starship.toml ~/.config/starship.toml
 
 # Icons
-ln -sT ~/dotfiles/icons ~/.local/share/icons
-
+create_symlink ~/dotfiles/icons ~/.local/share/icons
 
 # NodeJS
 ##
@@ -120,12 +113,8 @@ ln -sT ~/dotfiles/icons ~/.local/share/icons
 read -p "Press enter when nvm is installed"
 nvm install --lts
 
-
-
 # Zig
 sudo zypper install --no-recommends zig
-
-
 
 # Fonts
 ##
@@ -135,54 +124,38 @@ sudo zypper install --no-recommends zig
 ##
 read -p "Press enter when the fonts are installed"
 
-
-
 # Neovim
 sudo zypper install --no-recommends neovim
 ## Recommended
 sudo zypper install inotify-tools wl-clipboard
-## Config symlink
-ln -sT ~/dotfiles/nvim ~/.config/nvim
-
-
+create_symlink ~/dotfiles/nvim ~/.config/nvim
 
 # Neovide
 sudo zypper install neovide
 
-
 # Ripgrep
 sudo zypper install ripgrep
-
-
 
 # Fzf
 sudo zypper install fzf
 mkdir -p ~/.local/bin/fzf
-ln -sT ~/dotfiles/fzf/rose-pine.sh ~/.local/bin/fzf/rose-pine.sh
-
-
+create_symlink ~/dotfiles/fzf/rose-pine.sh ~/.local/bin/fzf/rose-pine.sh
 
 # Btop
 sudo zypper install btop
-ln -sT ~/dotfiles/btop/btop.conf ~/.config/btop/btop.conf
-ln -sT ~/dotfiles/btop/themes ~/.config/btop/themes
-
-
+create_symlink ~/dotfiles/btop/btop.conf ~/.config/btop/btop.conf
+create_symlink ~/dotfiles/btop/themes ~/.config/btop/themes
 
 # Neofetch/Fastfetch
 sudo zypper install fastfetch
 
-
-
 # Git
-ln -sT ~/dotfiles/gitconfig/opensuse/.gitconfig ~/.gitconfig-default
+create_symlink ~/dotfiles/gitconfig/opensuse/.gitconfig ~/.gitconfig-default
 if ! grep -q "~/.gitconfig-default" ~/.gitconfig; then
   echo "[include]" >> ~/.gitconfig
   echo "	path = ~/.gitconfig-default" >> ~/.gitconfig
 fi
 sudo zypper install git-filter-repo
-
-
 
 # Nvidia
 read -p "Do you want to install the NVIDIA drivers? (y/n): " install_nvidia
@@ -192,7 +165,6 @@ if [[ $install_nvidia == "y" ]]; then
   sudo zypper install --auto-agree-with-licenses nvidia-video-G06 nvidia-gl-G06 nvidia-compute-G06 nvidia-compute-utils-G06
 fi
 
-
 # Remove SDDM
 read -p "Do you want to remove SDDM? (y/n): " remove_sddm
 if [[ $remove_sddm == "y" ]]; then
@@ -200,12 +172,9 @@ if [[ $remove_sddm == "y" ]]; then
   sudo systemctl set-default multi-user.target
 fi
 
-
-
 # Hyprland autostart
 read -p "Do you want to set Hyprland as default? (y/n): " set_hyprland
 if [[ $set_hyprland == "y" ]]; then
-  # Check that marker is not already set in .profile
   if ! grep -q "# Hyprland autostart" ~/.profile; then
     echo "# Hyprland autostart" >> ~/.profile
     echo "if [ -z \"\$DISPLAY\" ] && [ \"\$(tty)\" = \"/dev/tty1\" ]; then" >> ~/.profile
@@ -214,9 +183,9 @@ if [[ $set_hyprland == "y" ]]; then
   fi
 fi
 
-
-# Uninstall unecessary packages
+# Uninstall unnecessary packages
 read -p "Do you want to uninstall xterm and related packages? (y/n): " uninstall_packages
 if [[ $uninstall_packages == "y" ]]; then
   sudo zypper remove xterm xterm-bin xtermset xterm-resize
 fi
+
